更新今天白天的联调内容：用户个人中心API的剩下3个接口

+++ back/src/main/java/com/bbt/back/dao/RecordDao.java
@@ -27,7 +27,7 @@ public interface RecordDao {
     //新增
     List<Record> selectByUserId(int userId);
 
-    Integer findRecordNumByUserId(Integer userId);
+    Long findRecordNumByUserId(Integer userId);
 
     Integer findProductNumByUserIdAndType(@Param("userId")Integer userId ,@Param("productType") Integer productType);
